requirements.txt
certifi @ file:///C:/b/abs_ac29jvt43w/croot/certifi_1665076682579/work/certifi
charset-normalizer==2.1.1
click==8.1.3
colorama==0.4.5
contourpy==1.0.5
cycler==0.11.0
Flask==2.2.2
fonttools==4.37.4
idna==3.4
importlib-metadata==5.0.0
imutils==0.5.4
itsdangerous==2.1.2
Jinja2==3.1.2
kiwisolver==1.4.4
MarkupSafe==2.1.1
matplotlib==3.6.1
numpy==1.23.3
opencv-python==4.6.0.66
packaging==21.3
pandas==1.5.0
Pillow==9.2.0
py-mon==1.1.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2022.4
requests==2.28.1
six==1.16.0
torch==1.12.1
torchvision==0.13.1
typing_extensions==4.4.0
urllib3==1.26.12
watchdog==2.1.9
Werkzeug==2.2.2
wincertstore==0.2
zipp==3.9.0
